fix lastAttack not being correct
