+++ module/item/SR5Item.ts
@@ -2,7 +2,6 @@ import { Helpers } from '../helpers';
 import { SR5Actor } from '../actor/SR5Actor';
 import { ShadowrunRollDialog } from '../apps/dialogs/ShadowrunRollDialog';
 import Template from '../template';
-import DeviceData = Shadowrun.DeviceData;
 import ModList = Shadowrun.ModList;
 import AttackData = Shadowrun.AttackData;
 import AttributeField = Shadowrun.AttributeField;
@@ -13,6 +12,7 @@ import SpellForceData = Shadowrun.SpellForceData;
 import ComplexFormLevelData = Shadowrun.ComplexFormLevelData;
 import FireRangeData = Shadowrun.FireRangeData;
 import BlastData = Shadowrun.BlastData;
+import { ChatData } from './ChatData';
 
 export class SR5Item extends Item {
     labels: {} = {};
@@ -23,19 +23,19 @@ export class SR5Item extends Item {
     getLastFireMode(): FireModeData | undefined {
         return this.getFlag('shadowrun5e', 'lastFireMode');
     }
-    setLastFireMode(fireMode: FireModeData) {
+    async setLastFireMode(fireMode: FireModeData) {
         return this.setFlag('shadowrun5e', 'lastFireMode', fireMode);
     }
     getLastSpellForce(): SpellForceData {
         return this.getFlag('shadowrun5e', 'lastSpellForce');
     }
-    setLastSpellForce(force: SpellForceData) {
+    async setLastSpellForce(force: SpellForceData) {
         return this.setFlag('shadowrun5e', 'lastSpellForce', force);
     }
     getLastComplexFormLevel(): ComplexFormLevelData {
         return this.getFlag('shadowrun5e', 'lastComplexFormLevel');
     }
-    setLastComplexFormLevel(level: ComplexFormLevelData) {
+    async setLastComplexFormLevel(level: ComplexFormLevelData) {
         return this.setFlag('shadowrun5e', 'lastComplexFormLevel', level);
     }
     getLastFireRange(): FireRangeData {
@@ -48,11 +48,10 @@ export class SR5Item extends Item {
     getLastAttack(): AttackData | undefined {
         return this.getFlag('shadowrun5e', 'lastAttack');
     }
-    setLastAttack(attack: AttackData) {
+    async setLastAttack(attack: AttackData) {
+        // unset the flag first to clear old data, data can get weird if not done
+        await this.unsetFlag('shadowrun5e', 'lastAttack');
         return this.setFlag('shadowrun5e', 'lastAttack', attack);
-    }
-    clearLastAttack() {
-        return this.unsetFlag('shadowrun5e', 'lastAttack');
     }
 
     async update(data, options?) {
@@ -186,10 +185,7 @@ export class SR5Item extends Item {
         item['properties'] = this.getChatData().properties;
     }
 
-    async roll(event) {
-        if (Helpers.hasModifiers(event)) {
-            return this.rollTest(event);
-        }
+    async postCard() {
         // we won't work if we don't have an actor
         if (!this.actor) return;
         const { token } = this.actor;
@@ -236,7 +232,8 @@ export class SR5Item extends Item {
         data.description.value = TextEditor.enrichHTML(data.description.value, htmlOptions);
 
         const props = [];
-        this[`_${this.data.type}ChatData`](duplicate(data), labels, props);
+        const func = ChatData[this.data.type];
+        if (func) func(duplicate(data), labels, props, this);
 
         data.properties = props.filter((p) => !!p);
 
@@ -260,300 +257,6 @@ export class SR5Item extends Item {
         const mod = this.getOpposedTestModifier();
         if (mod) name += ` ${mod}`;
         return name;
-    }
-
-    _ammoChatData(data, labels, props) {}
-
-    _modificationChatData(data, labels, props) {}
-
-    _actionChatData(data, labels, props) {
-        if (data.action) {
-            const labelStringList: string[] = [];
-            if (data.action.skill) {
-                labelStringList.push(Helpers.label(data.action.skill));
-                labelStringList.push(Helpers.label(data.action.attribute));
-            } else if (data.action.attribute2) {
-                labelStringList.push(Helpers.label(data.action.attribute));
-                labelStringList.push(Helpers.label(data.action.attribute2));
-            } else if (data.action.attribute) {
-                labelStringList.push(Helpers.label(data.action.attribute));
-            }
-            if (data.action.mod) {
-                labelStringList.push(`${game.i18n.localize('SR5.ItemMod')} (${data.action.mod})`);
-                // TODO when all mods are modlists
-                // Object.entries(data.action.mod).forEach(([key, value]) =>
-                //     labelStringList.push(`${game.i18n.localize(key)} (${value})`)
-                // );
-            }
-            if (labelStringList.length) {
-                labels.roll = labelStringList.join(' + ');
-            }
-
-            if (data.action.opposed.type) {
-                const { opposed } = data.action;
-                if (opposed.type !== 'custom')
-                    labels.opposedRoll = `vs. ${Helpers.label(opposed.type)}`;
-                else if (opposed.skill)
-                    labels.opposedRoll = `vs. ${Helpers.label(opposed.skill)}+${Helpers.label(
-                        opposed.attribute
-                    )}`;
-                else if (opposed.attribute2)
-                    labels.opposedRoll = `vs. ${Helpers.label(opposed.attribute)}+${Helpers.label(
-                        opposed.attribute2
-                    )}`;
-                else if (opposed.attribute)
-                    labels.opposedRoll = `vs. ${Helpers.label(opposed.attribute)}`;
-            }
-
-            // setup action props
-            // go in order of "Limit/Accuracy" "Damage" "AP"
-            // don't add action type if set to 'varies' or 'none' as that's pretty much useless info
-            if (
-                data.action.type !== '' &&
-                data.action.type !== 'varies' &&
-                data.action.type !== 'none'
-            ) {
-                props.push(`${Helpers.label(data.action.type)} Action`);
-            }
-            if (data.action.limit.value) props.push(`Limit ${data.action.limit.value}`);
-            if (data.action.damage.type.value) {
-                const { damage } = data.action;
-                let damageString = '';
-                let elementString = '';
-                if (damage.value) {
-                    damageString = `DV ${damage.value}${
-                        damage.type.value ? damage.type.value.toUpperCase().charAt(0) : ''
-                    }`;
-                }
-                if (damage.element.value) {
-                    // if we have a damage value and are electric, follow the convention of (e) after
-                    if (damage.value) {
-                        if (damage.element.value === 'electricity') {
-                            damageString += ' (e)';
-                        } else {
-                            elementString = Helpers.label(damage.element.value);
-                        }
-                    } else {
-                        elementString = Helpers.label(damage.element.value);
-                    }
-                }
-                if (damageString) props.push(damageString);
-                if (elementString) props.push(elementString);
-                if (damage.ap && damage.ap.value) props.push(`AP ${damage.ap.value}`);
-            }
-        }
-    }
-
-    _sinChatData(data, labels, props) {
-        props.push(`Rating ${data.technology.rating}`);
-        data.licenses.forEach((license) => {
-            props.push(`${license.name} R${license.rtg}`);
-        });
-    }
-
-    _contactChatData(data, labels, props) {
-        props.push(data.type);
-        props.push(`Connection ${data.connection}`);
-        props.push(`Loyalty ${data.loyalty}`);
-    }
-
-    _lifestyleChatData(data, labels, props) {
-        props.push(Helpers.label(data.type));
-        if (data.cost) props.push(`¥${data.cost}`);
-        if (data.comforts) props.push(`Comforts ${data.comforts}`);
-        if (data.security) props.push(`Security ${data.security}`);
-        if (data.neighborhood) props.push(`Neighborhood ${data.neighborhood}`);
-        if (data.guests) props.push(`Guests ${data.guests}`);
-    }
-
-    _adept_powerChatData(data, labels, props) {
-        this._actionChatData(data, labels, props);
-        props.push(`PP ${data.pp}`);
-        props.push(Helpers.label(data.type));
-    }
-
-    _armorChatData(data, labels, props) {
-        if (data.armor) {
-            if (data.armor.value)
-                props.push(`Armor ${data.armor.mod ? '+' : ''}${data.armor.value}`);
-            if (data.armor.acid) props.push(`Acid ${data.armor.acid}`);
-            if (data.armor.cold) props.push(`Cold ${data.armor.cold}`);
-            if (data.armor.fire) props.push(`Fire ${data.armor.fire}`);
-            if (data.armor.electricity) props.push(`Electricity ${data.armor.electricity}`);
-            if (data.armor.radiation) props.push(`Radiation ${data.armor.radiation}`);
-        }
-    }
-
-    _complex_formChatData(data, labels, props) {
-        this._actionChatData(data, labels, props);
-        props.push(Helpers.label(data.target), Helpers.label(data.duration));
-        const { fade } = data;
-        if (fade > 0) props.push(`Fade L+${fade}`);
-        else if (fade < 0) props.push(`Fade L${fade}`);
-        else props.push('Fade L');
-    }
-
-    _cyberwareChatData(data, labels, props) {
-        this._actionChatData(data, labels, props);
-        this._armorChatData(data, labels, props);
-        if (data.essence) props.push(`Ess ${data.essence}`);
-    }
-
-    _deviceChatData(data: DeviceData, labels, props) {
-        if (data.technology && data.technology.rating)
-            props.push(`Rating ${data.technology.rating}`);
-        if (data.category === 'cyberdeck') {
-            for (const attN of Object.values(data.atts)) {
-                props.push(`${Helpers.label(attN.att)} ${attN.value}`);
-            }
-        }
-    }
-
-    _equipmentChatData(data, labels, props) {
-        if (data.technology && data.technology.rating)
-            props.push(`Rating ${data.technology.rating}`);
-    }
-
-    _qualityChatData(data, labels, props) {
-        this._actionChatData(data, labels, props);
-        props.push(Helpers.label(data.type));
-    }
-
-    // add properties for spell data, follow order in book
-    _spellChatData(data, labels, props) {
-        // first category and type
-        props.push(Helpers.label(data.category), Helpers.label(data.type));
-
-        // add subtype tags
-        if (data.category === 'combat') {
-            props.push(Helpers.label(data.combat.type));
-        } else if (data.category === 'health') {
-        } else if (data.category === 'illusion') {
-            props.push(data.illusion.type);
-            props.push(data.illusion.sense);
-        } else if (data.category === 'manipulation') {
-            if (data.manipulation.damaging) props.push('Damaging');
-            if (data.manipulation.mental) props.push('Mental');
-            if (data.manipulation.environmental) props.push('Environmental');
-            if (data.manipulation.physical) props.push('Physical');
-        } else if (data.category === 'detection') {
-            props.push(data.illusion.type);
-            props.push(data.illusion.passive ? 'Passive' : 'Active');
-            if (data.illusion.extended) props.push('Extended');
-        }
-        // add range
-        props.push(Helpers.label(data.range));
-
-        // add action data
-        this._actionChatData(data, labels, props);
-
-        // add duration data
-        props.push(Helpers.label(data.duration));
-
-        // add drain data
-        const { drain } = data;
-        if (drain > 0) props.push(`Drain F+${drain}`);
-        else if (drain < 0) props.push(`Drain F${drain}`);
-        else props.push('Drain F');
-
-        labels.roll = 'Cast';
-    }
-
-    _weaponChatData(data, labels, props) {
-        this._actionChatData(data, labels, props);
-        for (let i = 0; i < props.length; i++) {
-            const prop = props[i];
-            if (prop.includes('Limit')) {
-                props[i] = prop.replace('Limit', 'Accuracy');
-            }
-        }
-
-        const equippedAmmo = this.getEquippedAmmo();
-        if (equippedAmmo && data.ammo && data.ammo.current?.max) {
-            if (equippedAmmo) {
-                const { current, spare_clips } = data.ammo;
-                if (equippedAmmo.name)
-                    props.push(`${equippedAmmo.name} (${current.value}/${current.max})`);
-                if (equippedAmmo.data.data.blast.radius)
-                    props.push(
-                        `${game.i18n.localize('SR5.BlastRadius')} ${
-                            equippedAmmo.data.data.blast.radius
-                        }m`
-                    );
-                if (equippedAmmo.data.data.blast.dropoff)
-                    props.push(
-                        `${game.i18n.localize('SR5.Dropoff')} ${
-                            equippedAmmo.data.data.blast.dropoff
-                        }/m`
-                    );
-                if (spare_clips && spare_clips.max)
-                    props.push(
-                        `${game.i18n.localize('SR5.SpareClips')} (${spare_clips.value}/${
-                            spare_clips.max
-                        })`
-                    );
-            }
-        }
-
-        if (data.technology?.conceal?.value) {
-            props.push(`${game.i18n.localize('SR5.Conceal')} ${data.technology.conceal.value}`);
-        }
-
-        if (data.category === 'range') {
-            if (data.range.rc) {
-                let rcString = `${game.i18n.localize('SR5.RecoilCompensation')} ${
-                    data.range.rc.value
-                }`;
-                if (this.actor) {
-                    rcString += ` (${game.i18n.localize('SR5.Total')} ${
-                        this.actor.data.data.recoil_compensation + data.range.rc.value
-                    })`;
-                }
-                props.push(rcString);
-            }
-            if (data.range.modes) {
-                const newModes: string[] = [];
-                const { modes } = data.range;
-                if (modes.single_shot) newModes.push('SR5.WeaponModeSingleShotShort');
-                if (modes.semi_auto) newModes.push('SR5.WeaponModeSemiAutoShort');
-                if (modes.burst_fire) newModes.push('SR5.WeaponModeBurstFireShort');
-                if (modes.full_auto) newModes.push('SR5.WeaponModeFullAutoShort');
-                props.push(newModes.map((m) => game.i18n.localize(m)).join('/'));
-            }
-            if (data.range.ranges)
-                props.push(Array.from(Object.values(data.range.ranges)).join('/'));
-        } else if (data.category === 'melee') {
-            if (data.melee.reach) {
-                const reachString = `${game.i18n.localize('SR5.Reach')} ${data.melee.reach}`;
-                // find accuracy in props and insert ourselves after it
-                const accIndex = props.findIndex((p) => p.includes('Accuracy'));
-                if (accIndex > -1) {
-                    props.splice(accIndex + 1, 0, reachString);
-                } else {
-                    props.push(reachString);
-                }
-            }
-        } else if (data.category === 'thrown') {
-            const { blast } = data.thrown;
-            if (blast?.radius)
-                props.push(`${game.i18n.localize('SR5.BlastRadius')} ${blast.radius}m`);
-            if (blast?.dropoff)
-                props.push(`${game.i18n.localize('SR5.Dropoff')} ${blast.dropoff}/m`);
-
-            if (data.thrown.ranges) {
-                const mult =
-                    data.thrown.ranges.attribute && this.actor
-                        ? this.actor.data.data.attributes[data.thrown.ranges.attribute].value
-                        : 1;
-                const ranges = [
-                    data.thrown.ranges.short,
-                    data.thrown.ranges.medium,
-                    data.thrown.ranges.long,
-                    data.thrown.ranges.extreme,
-                ];
-                props.push(ranges.map((v) => v * mult).join('/'));
-            }
-        }
     }
 
     getEquippedAmmo() {
@@ -704,6 +407,11 @@ export class SR5Item extends Item {
         };
 
         const lastAttack = this.getLastAttack();
+        console.log('opp');
+        console.log('');
+        console.log(lastAttack);
+        console.log('');
+        console.log('');
         if (lastAttack) {
             options['incomingAttack'] = lastAttack;
             options.cover = true;
@@ -994,6 +702,11 @@ export class SR5Item extends Item {
 
         const blastData = this.getBlastData();
         if (blastData) data.blast = blastData;
+
+        console.log('atkData');
+        console.log('');
+        console.log(data);
+        console.log('');
 
         return data;
     }
