+++ templates/rolls/item-card.html
@@ -10,8 +10,6 @@
 
   <div class="card-buttons">
     {{#if hasRoll}}<button data-action='roll'>{{labels.roll}}</button>{{/if}}
-    {{#if hasOpposedRoll}}<button data-action='opposed-roll'>{{labels.opposedRoll}}</button>{{/if}}
-    {{#if hasTemplate}}<button data-action='place-template'>{{localize 'SR5.PlaceTemplate'}}</button>{{/if}}
   </div>
 
   <footer class='card-footer'>
