+++ templates/rolls/roll-card.html
@@ -76,7 +76,7 @@
             <div class="damage-line">
                 {{#if soak.value}}
                 <div>
-                    <span class="key">{{localize "SR5.Total"}} {{localize "SR5.DV"}}: </span>
+                    <span class="key">{{localize "SR5.Incoming"}} {{localize "SR5.DV"}}: </span>
                     <span class="value">{{soak.value}}{{damageAbbreviation soak.type.value}}</span>
                 </div>
                 {{/if}} {{#if soak.element.value}}
